+++ CHANGES.md
@@ -4,6 +4,10 @@ moodle-local_boostnavigation
 Changes
 -------
 
+### Unreleased
+
+* 2019-04-12 - Fix a wrong node example in the language pack.
+
 ### v3.6-r3
 
 * 2019-03-29 - Bugfix: Remove debugging notices about undefined properties when running unit tests - Credits to G El-Zoghbi.